+++ aws/fetch/manual_fetchers.go
@@ -1412,3 +1412,6 @@ func addManualKinesisFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
 
 func addManualRedshiftFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
 }
+
+func addManualCodepipelineFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
+}