+++ aws/services/init.go
@@ -27,9 +27,9 @@ import (
 )
 
 var (
-	AccessService, InfraService, StorageService, MessagingService, DNSService, LambdaService, MonitoringService, CdnService, CloudformationService cloud.Service
-	EksService, DynamodbService, SecretsmanagerService, ApigatewayService, SsmService, EfsService, CloudtrailService, CloudwatchlogsService        cloud.Service
-	ElasticacheService, EventbridgeService, StepfunctionsService, WafService, ConfigserviceService, KinesisService, RedshiftService                cloud.Service
+	AccessService, InfraService, StorageService, MessagingService, DNSService, LambdaService, MonitoringService, CdnService, CloudformationService       cloud.Service
+	EksService, DynamodbService, SecretsmanagerService, ApigatewayService, SsmService, EfsService, CloudtrailService, CloudwatchlogsService              cloud.Service
+	ElasticacheService, EventbridgeService, StepfunctionsService, WafService, ConfigserviceService, KinesisService, RedshiftService, CodepipelineService cloud.Service
 )
 
 func Init(profile, region string, extraConf map[string]any, log *logger.Logger, profileSetterCallback func(val string) error, enableNetworkMonitor bool) error {
@@ -69,6 +69,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	ConfigserviceService = NewConfigservice(cfg, profile, extraConf, log)
 	KinesisService = NewKinesis(cfg, profile, extraConf, log)
 	RedshiftService = NewRedshift(cfg, profile, extraConf, log)
+	CodepipelineService = NewCodepipeline(cfg, profile, extraConf, log)
 
 	cloud.ServiceRegistry[InfraService.Name()] = InfraService
 	cloud.ServiceRegistry[AccessService.Name()] = AccessService
@@ -94,6 +95,7 @@ func Init(profile, region string, extraConf map[string]any, log *logger.Logger, 
 	cloud.ServiceRegistry[ConfigserviceService.Name()] = ConfigserviceService
 	cloud.ServiceRegistry[KinesisService.Name()] = KinesisService
 	cloud.ServiceRegistry[RedshiftService.Name()] = RedshiftService
+	cloud.ServiceRegistry[CodepipelineService.Name()] = CodepipelineService
 
 	awsspec.CommandFactory = &awsspec.AWSFactory{
 		Log: log,